refactor: move detour.ts into models
